requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.3
cssselect2==0.7.0
Django==4.1.10
django-appconf==1.0.5
django-cors-headers==4.3.1
django-htmx==1.16.0
django-js-asset==2.1.0
django-mptt==0.15.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-select2==8.1.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
et-xmlfile==1.1.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
lxml==4.9.3
mssql-django==1.2
mypy-extensions==1.0.0
numpy==1.25.2
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
phonenumbers==8.13.19
Pillow==10.0.0
platformdirs==3.10.0
pycparser==2.21
pyHanko==0.20.0
pyhanko-certvalidator==0.23.0
pyodbc==4.0.39
pypdf==3.16.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.0.4
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
xlwt==1.3.0
